+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## In development
 
+- Upgrade grpc-gateway and add required protobuf options [#204](https://github.com/nre-learning/antidote-core/pull/204)
 - Add (deprecated) 'legacy' image flavor [#202](https://github.com/nre-learning/antidote-core/pull/202)
 - Move networkpolicy creation to the beginning of the creation of livelesson resources  [#201](https://github.com/nre-learning/antidote-core/pull/201)
 - Fix race bug in job completion check [#200](https://github.com/nre-learning/antidote-core/pull/200)